LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-inline-004-expected.txt

spacer
0 12 a1 34
0 12 a1 a1 a1 345
0 12 a1
a1 345
0 12 a1 a1 a1 345

PASS .target 1
PASS .target 2
PASS .target 3
PASS .target 4
PASS .target 5
PASS .target 6
PASS .target 7
PASS .target 8
PASS .target 9
PASS .target 10
PASS .target 11
PASS .target 12
PASS .target 13
PASS .target 14
PASS .target 15
PASS .target 16
PASS .target 17
PASS .target 18
FAIL .target 19 assert_equals:
<span class="target target1-pos" data-offset-x="30" data-offset-y="-40" data-expected-width="80" data-expected-height="50"></span>
width expected 80 but got 210
FAIL .target 20 assert_equals:
<span class="target target1-size" data-offset-x="30" data-offset-y="-40" data-expected-width="80" data-expected-height="50"></span>
width expected 80 but got 100
FAIL .target 21 assert_equals:
<span class="target target1-pos" data-offset-x="160" data-offset-y="0" data-expected-width="80" data-expected-height="50"></span>
width expected 80 but got 210
FAIL .target 22 assert_equals:
<span class="target target1-size" data-offset-x="160" data-offset-y="0" data-expected-width="80" data-expected-height="50"></span>
width expected 80 but got 100
FAIL .target 23 assert_equals:
<span class="target target1-pos" data-offset-x="160" data-offset-y="0" data-expected-width="80" data-expected-height="50"></span>
width expected 80 but got 210
FAIL .target 24 assert_equals:
<span class="target target1-size" data-offset-x="160" data-offset-y="0" data-expected-width="80" data-expected-height="50"></span>
width expected 80 but got 100
